008 - MCP Demo/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
openai
mcp